+++ sdk/core/Azure.Core/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.3.0-preview.1 (Unreleased)
 
+
+## 1.2.2 (2020-06-04)
+
 ### Bugfix
 - Retry server timeouts on .NET Framework.