+++ sdk/identity/Azure.Identity/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.0-preview.4 (Unreleased)
 
 ### New Features
+- Makes `AzureCliCredential`, `VisualStudioCredential` and `VisualStudioCodeCredential` public to allow direct usage.
 - Added `Authenticate` methods to `UsernamePasswordCredential`
 
 ## 1.2.0-preview.3